Trac #28328: use a smaller category for unital algebra base ring morphism.
The morphism that takes a base ring element into a unital algebra
needs some category to live in. When the algebra is not necessarily
associative, the category of unital magmas was used as the result of
some other implementation detail. However, we can strenghthen that a
bit: the only requirements we have are that both the base ring and the
associative algebra need to live in the category. Thus this commit
shrinks the category to those unital magmas that also happen to be
commutative additive groups with a distributive law.