test if a parent represents real/complex numbers
The de facto standard way of checking if a parent represents a subring
of the reals is currently something like:

  RealField(mpfr_prec_min()).has_coerce_map_from(P)

(and similarly for complex numbers). We introduce methods that one can
call instead, with a default implementation similar to the above, and
some tooling to handle the case where P is a type.