28574: construct Macaulay2 polynomial rings over arbitrary base rings
This simplifies the construction of Macaulay2 polynomial rings by
constructing the base ring in Macaulay2 first.

We add a special case in the conversion of Galois fields, since in
Macaulay2 it is more natural to define finite fields as `ZZ/p` rather
than `GF p`. This has a positive effect on performance.