Trac #28855: always return PolynomialRing elements from random_element().
The multivariate polynomial ring's random_element() method contains a
short-circuit path for when the user requests a degree-zero element.
Instead of returning a polynomial ring element, it returns a base ring
element -- which are more or less the same thing -- but not quite.

Getting a base ring element when you expect a polynomial can be
annoying, and forces you to add special cases to algorithms that
invoke polynomial methods. This commit wraps the return value (the
base ring element) in a call to self(), ensuring that it has the
correct parent. Two new tests for the degree-zero and general cases
were also added to check that random elements have the right parent().