Trac #28905: add header/library checks for cddlib.
A few consumers use cddlib as a library, so we need to ensure that
its headers can be found and used. There is some confusion about
where the headers should be located, but this is "solved" by the
fact that latte and gfan only look for them in the top-level path
like /usr/include/cdd.h. The new check makes sure the headers are
where those packages expect them to be, and ensures that we can link
to cddlib.