Trac #29245: add "distros" directory with an entry for Gentoo.
Since Gentoo ships this package as sci-libs/fplll, we add a file named
gentoo.txt containing "sci-libs/fplll" to the new distros directory.
This information is useful to users who wish to install the system
package, but who may not know what the package is called on the
various distributions.