Trac #29245: add Debian package information for fplll.
On Debian, one needs the libfplll-dev package to link programs (like
SageMath) against fplll. This commit adds a new file debian.txt to the
build/pkgs/fplll/distros directory that mentions that package along
with the fplll-tools package that Debian provides. The latter is small
and can't hurt, even though it may not strictly be necessary.