Trac #29297: add special case to is_unitary() for trivial RDF/CDF matrices.
The default is_unitary() algorithm for RDF and CDF matrices doesn't
know how to handle a trivial (zero-by-zero) matrix. The exact error
message that results from it crashing is doctested, and in fact that
message is somewhat variable depdending on the libraries installed on
the system, which can lead to test failures.

Rather than make the test in question more robust, this commit fixes
the underlying problem by adding a special case for trivial RDF/CDF
matrices to make is_unitary() act like it does on other rings. Namely,
it now always returns True.