build/pkgs/openblas/spkg-configure.m4: Do not use AC_FC_FUNC.
It was wrapped in AC_FC_FREEFORM so that there is no error when no Fortran compiler is found.
But AC_FC_FREEFORM is AC_DEFUN_ONCE, and is invoked by gfortran/spkg-configure.m4 as well.
So this line will actually not be executed.

Instead we check all possible Fortran name manglings by hand.