Trac #29369: factor out brial's python module into a sage_brial package.
Sage's "brial" package was actually installing two separate packages,
the C++ library, and a python module called "sage-brial". As a result,
detecting (and using) only the C++ library on the system was
insufficient to replace the SPKG.

Ultimately, sage-brial provides only sage-specific glue to the brial
library, and should be incorporated into the sage library itself. But
for now, this commit splits out the python module into another package
called "sage_brial", which of course depends on "brial" to supply the
C++ library. This is the path of least resistance, and allows us to
use the C++ library from the system if it is present.