Trac #29458: new spkg-configure.m4 for libbraiding.
This commit adds an spkg-configure.m4 file to detect a system copy
of libbraiding. Since it's a C++ library and doesn't ship a pkg-config
file, we have to compile, link, and run a test program to make sure
that it works. I've taken a short example that can be checked in Sage,
and have converted it to C++ for this purpose.