Trac #29454: hack m4/ppl.m4 to define PPL_LIBS.
The upstream AM_PATH_PPL macro gets its LDFLAGS from the ppl-config
program, but fails to differentiate between LDFLAGS and LIBS. This
ultimately causes "-lppl" to be appended to the command-line in the
wrong spot. This commit hacks the macro to additionally define a
PPL_LIBS variable in the event that PPL is detected. This allows us to
append the flag in the right place, and avoid appending it in the
wrong place.