Trac #29453: require at least fflas-ffpack-2.4.0.
The version of linbox (1.6.3) that currently ships with SageMath
needs fflas-ffpack-2.4.0 or newer. Since we can't use linbox from
the system at the moment, that means that SageMath depends on the
same version of fflas-ffpack to be able to build. This commit updates
the lower bound from fflas-ffpack-2.3.2 to fflas-ffpack-2.4.0. This
should be revisited when we are able to use the system's linbox.