Trac #29491: refactor gp2c's spkg-configure.m4.
Thanks to recent improvements in the build system, we no longer need
to manually check if pari (a dependency of gp2c) is installed; nor do
we need to alert the user to the possibility of installing the gp2c
system package. This commit refactors gp2c's SAGE_SPKG_CONFIGURE to
avoid doing those things.

The refactoring also eliminates a costly "find /usr -name pari.cfg"
command that was being run during every invocation of the ./configure
script where the system pari was used. On systems where /usr is large,
enumerating its contents can take forever. Instead, we now check a
fixed(ish) set of locations where pari.cfg is known to be installed on
distributions we support. This may prove slightly less robust than
"find", but it's easy to add new locations in the future, and doing so
avoids penalizing the common case like the "find" command does.