Add logarithmic point selection to generate_plot_points()
In plot(), if scale == 'semilogx' or 'loglog' when calling plot, it will
now call the modified mode of generate_plot_points() that will select
the points to evaluate exponentially.  This places the evaluated points
evenly along the viewport's x axis which decreases the number of points
that are needed for accurate log-scale plots.