Fix creating valuations from prime ideals
The initial versions of the valuations coming out of
mac_lane_approximants() might not be able to detect the generators of
the defining prime ideal depending on how those generators were chosen.
However, eventually the correct refinement must have v(g)>0 for all
generators. Since v'(g) ≥ v(g) for a refinement v' of v, we are also
sure not to detect the wrong extension in the initial steps.