Trac #29345: fix some bashisms in sage-env's resolvelinks() function.
This function was using bash-specific parameter expansions to find-
and-replace substrings. Fortunately, most of the substrings were at
the beginning of the big string, where the POSIX shortest-prefix
strip, followed by a prepend, accomplishes the same thing. The other
instance was replacing multiple slashes by a single slash in a path,
which is now implemented with a call to sed.