#28517 bug in real_nth_root() revealed by rebasing this ticket
A code path apparently intended for the case real_nth_root(integer,
integer) had its result silently dropped, leading to a fallback being
used. This code path should also have been the one taken when the first
argument is an interval.

We fix that and move/update some examples and tests to better reflect
the role of each method.