Trac #30533: hide automake (stderr) output from ./bootstrap -q.
Eventually our ./bootstrap process runs "automake", which outputs a
bunch of lines like

  configure.ac:89: installing 'config/config.guess'
  configure.ac:89: installing 'config/config.sub'
  ...

to stderr. With the recent addition of the "-q" (quiet) flag to the
bootstrap script, it would be nice if we could hide that expected
output. This commit does so, but in a way that makes me feel dirty.
The goal is to filter only those messages, and only from stderr when
automake is run, and then to re-output to stderr -- and not stdout,
like "sed" normally would. I have managed to achieve this using only
POSIX constructs... but what's happening is not at all self-explanatory.
And piping the result of automake to sed clobbers the return value of
automake in the "automake && autoconf" chain.

Running autoconf after automake has failed is not too big of a deal,
however. We proceed if autoconf fails anyway; the script is not very
picky about its error handling.