Trac #30605: update polyhedral cone containment documentation.
This commit adds a few somewhat redundant tests to the _contains()
method of the ConvexRationalPolyhedralCone class to ensure that it can
be used on irrational and complex inputs. Many of these are indirectly
tested by cone._ambient_space_point(), but the fact that cone
containment uses that particular function is an implementation detail,
so it's nice to see the containment properties tested directly.

In the process, the _contains() method's docstring was reorganized,
and a warning was added about using it to distinguish boundary points
over inexact rings.