Trac #30654: add fast inverse method for positive-definite matrices.
Positive-definite matrices can be inverted quickly and in a
numerically-stable way by first factoring the matrix and then
inverting the factors. This commit adds a new matrix method called
inverse_positive_definite() that uses the indefinite_factorization().

This will simplify the computation of maximal and critical angles
within polyhedral convex cones, in Trac ##29169.