upgrade ubuntu system in docker image
this fixes an issue with line endings in CI. The old version of
git on ubuntu xenial does not understand our .gitattributes or it does
not understand the .git directory that we copy into the docker build
context from a much newer version of git. Anyway, the broken line
endings are gone now.