Trac #31453: support generators in CFM _cartesian_product_of_elements().
The _cartesian_product_of_elements() method is supposed to support any
iterable, including a generator. However the implementation in the
CombinatorialFreeModule class indexes into its argument directly, and
does not work when that argument is a generator as opposed to, say, a
list. This commit alters that implementation a bit to make it support
generators.