Trac #29617: check for ecl-config instead of ecl.
The plan for ecl's ./configure check is to build a test program using
a header that defines its version, and to check that version against
the required one. If those headers are installed in an uncommon
location, the ecl-config program exists to provide the appropriate
include path(s). Since we will need those include path(s) anyway, we
might as well check for ecl-config instead of ecl itself.