Trac #29617: remove failing ECL test.
There is a test in sage/interfaces/tests.py that checks whether or not
the "ecl" executable handles stdout and stderr being "full" at the
same time. This fails without a custom patch, and therefore with
everyone's system copy of ecl.

This commit removes the test: running "ecl" on the command-line is not
a special sage interface to ecl. This is an upstream ecl bug that will
someday get fixed and be added to the ecl test suite. In the meantime,
it has no ill effects in sage besides the failing test.