Trac #31619: support is_positive_definite() over symbolic ring.
The is_positive_definite() method for matrices implicitly rejects
symbolic matrices. There's no principled way to deal with SR, but it's
still nice for inexperienced users when things "just work."  For
example, I think we should be able to say that a square diagonal
matrix with all "pi" entries on the diagonal is positive-definite.

This commits updates the real/complex check in is_positive_definite()
to explicitly allow SR, just in case the result will ultimately make
sense.