Trac #31619: check reality/complexity in is_positive_semidefinite().
One of the tests for is_positive_definite() shows that misleading
results can be returned over e.g. finite fields if we don't check
that the base ring of our matrix is real or complex first. We copy
that check into is_positive_semidefinite(), too.

We also fix a few tests for number fields that were missing a complex
embedding.